Records (#156)
* refactor: records container

abstract all query logic to hooks
records query functions
