+++ src/components/records/S2SRecord.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { RecordComponentProps } from '../../config';
-import { NetConfig, NoNullFields, S2SHistoryRecord } from '../../model';
+import { S2SHistoryRecord } from '../../model';
 import { convertToSS58, getNetworkMode } from '../../utils';
 import { CrosseState, ProgressDetail } from './ProgressDetail';
 import { Record } from './Record';
@@ -14,28 +14,26 @@ import { Record } from './Record';
  * FIXME: Can not find the corresponding event on target chain, because the message_id is missing;
  */
 // eslint-disable-next-line complexity
-export function S2SRecord({
-  record,
-  network: departure,
-  arrival,
-}: NoNullFields<RecordComponentProps<S2SHistoryRecord> & { arrival: NetConfig }>) {
+export function S2SRecord({ record, departure, arrival }: RecordComponentProps<S2SHistoryRecord>) {
   const { requestTxHash, responseTxHash, amount, result, endTimestamp, startTimestamp, recipient } = record;
   const step = result === 0 ? CrosseState.pending : CrosseState.claimed;
-  const isRedeem = useMemo(() => getNetworkMode(departure) === 'dvm', [departure]);
+  const isRedeem = useMemo(() => departure && getNetworkMode(departure) === 'dvm', [departure]);
+  const fromNetwork = useMemo(() => departure?.name || 'pangoro', [departure?.name]);
+  const toNetwork = useMemo(() => arrival?.name || 'pangoro', [arrival?.name]);
 
   return (
     <Record
-      from={{ network: departure.name, txHash: requestTxHash }}
-      to={{ network: arrival.name, txHash: requestTxHash }}
+      from={{ network: fromNetwork, txHash: requestTxHash }}
+      to={{ network: toNetwork, txHash: requestTxHash }}
       hasRelay={false}
       step={step}
       blockTimestamp={+(endTimestamp || startTimestamp || Date.now())}
-      recipient={isRedeem ? convertToSS58(recipient, arrival.ss58Prefix) : recipient}
+      recipient={isRedeem ? convertToSS58(recipient, arrival?.ss58Prefix ?? null) : recipient}
       assets={[{ amount, currency: isRedeem ? 'xORING' : 'ORING', unit: 'gwei' }]}
     >
       <ProgressDetail
-        from={{ network: departure.name, txHash: requestTxHash }}
-        to={{ network: departure.name, txHash: responseTxHash }}
+        from={{ network: fromNetwork, txHash: requestTxHash }}
+        to={{ network: fromNetwork, txHash: responseTxHash }}
         hasRelay={false}
         step={step}
         stepDescriptions={{
