+++ src/pages/CrossHistory.tsx
@@ -1,10 +1,10 @@
 import { RouteComponentProps } from 'react-router-dom';
-import { Records } from '../components/records/Records';
+import { CrossRecords } from '../components/records/CrossRecords';
 
 export function HistoryRecords(_: RouteComponentProps) {
   return (
     <div id="history-records" className="w-full mx-auto max-w-6xl relative">
-      <Records></Records>
+      <CrossRecords></CrossRecords>
     </div>
   );
 }
